Better names for the routes and define DashboardController
